# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ZenFM API
  version: 0.1.0
  description: |
    The local, single-owner ZenFM HTTP API. Browser requests authenticate with
    an opaque `zenfm_session` cookie and send the session-bound value returned
    as `csrfToken` in `X-ZenFM-CSRF` for every mutation. Personal API tokens use
    `Authorization: Bearer`; their restricted scope excludes owner security,
    token management, update, and lifecycle operations.
servers:
  - url: /
paths:
...
  /api/v1/files/raw:
    get:
      operationId: readFile
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
      responses:
        '200':
          description: Streamed regular-file content with range support.
          headers:
            Accept-Ranges:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/shares:
    get:
      operationId: listShares
      security:
        - sessionCookie: []
        - personalToken: []
      responses:
        '200':
          description: Active owner shares without password hashes or raw secret digests.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Share'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createShare
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShare'
      responses:
        '201':
          description: Share created; its capability URL is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Share'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/shares/{shareId}:
    get:
      operationId: getShare
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/ShareId'
      responses:
        '200':
          description: One share.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Share'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: revokeShare
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/CsrfOptionalForBearer'
        - $ref: '#/components/parameters/ShareId'
      responses:
        '204':
          description: Share and its public sessions revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/public/shares/{secret}:
    get:
      operationId: inspectPublicShare
      security: []
      parameters:
        - $ref: '#/components/parameters/ShareSecret'
        - $ref: '#/components/parameters/PublicPath'
      responses:
        '200':
          description: Public metadata, or a password challenge without secret details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicShare'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      operationId: unlockPublicShare
      security: []
      parameters:
        - $ref: '#/components/parameters/ShareSecret'
        - $ref: '#/components/parameters/PublicPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password]
              properties:
                password:
                  type: string
                  maxLength: 1024
      responses:
        '200':
          description: A scoped, 30-minute public-share session was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicShare'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UniformLoginFailure'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: zenfm_session
    personalToken:
      type: http
      scheme: bearer
      bearerFormat: opaque ZenFM personal token
  parameters:
...
    CsrfOptionalForBearer:
      in: header
      name: X-ZenFM-CSRF
      required: false
      schema:
        type: string
      description: Required for cookie-authenticated mutations; ignored for bearer authentication.
    Path:
      in: query
      name: path
      required: true
      schema:
        $ref: '#/components/schemas/Path'
...
    PublicPath:
      in: query
      name: path
      required: false
      description: Capability-relative descendant path. `/` addresses the share root and never reveals its owner path.
      schema:
        $ref: '#/components/schemas/Path'
...
    ShareId:
      in: path
      name: shareId
      required: true
      schema:
        $ref: '#/components/schemas/OpaqueId'
    ShareSecret:
      in: path
      name: secret
      required: true
      schema:
        type: string
        minLength: 22
        maxLength: 128
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing, expired, or revoked credentials.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UniformLoginFailure:
      description: Credentials were not accepted. Account existence is not disclosed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated principal lacks access, setup is incomplete, CSRF failed, or origin was rejected.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    RateLimited:
      description: Authentication or public-share attempt limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Path:
      type: string
      minLength: 1
      maxLength: 4096
      pattern: '^/'
      examples: [/Books]
    OpaqueId:
      type: string
      minLength: 16
      maxLength: 128
      pattern: '^[A-Za-z0-9_-]+$'
...
    FileEntry:
      type: object
      additionalProperties: false
      required: [name, path, type, size, modifiedAt]
      properties:
        name:
          type: string
        path:
          $ref: '#/components/schemas/Path'
        type:
          type: string
          enum: [file, directory, symlink, special]
        size:
          type: integer
          minimum: 0
        modifiedAt:
          type: string
          format: date-time
        mimeType:
          type: string
        hidden:
          type: boolean
        writable:
          type: boolean
...
    CreateShare:
      type: object
      additionalProperties: false
      required: [path]
      properties:
        path:
          $ref: '#/components/schemas/Path'
        name:
          type: string
          maxLength: 200
        password:
          type: string
          maxLength: 1024
        expiresInSeconds:
          type: integer
          minimum: 60
          maximum: 31536000
    Share:
      type: object
      additionalProperties: false
      required: [id, path, passwordProtected, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/OpaqueId'
        path:
          $ref: '#/components/schemas/Path'
        name:
          type: string
        url:
          type: string
        expiresAt:
          type: [string, 'null']
          format: date-time
        passwordProtected:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PublicShare:
      type: object
      additionalProperties: false
      required: [name, path]
      properties:
        name:
          type: string
        path:
          $ref: '#/components/schemas/Path'
        entry:
          $ref: '#/components/schemas/FileEntry'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/FileEntry'
        passwordRequired:
          type: boolean
        expiresAt:
          type: [string, 'null']
          format: date-time
...
    Problem:
      type: object
      additionalProperties: true
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference